Kapt: Do not use awkward ('$' -> '/') name mangling logic for KAPT3 class builder mode (partially fixes KT-19433)
Before this commit, internal names for nested classes were written as test/Foo/Bar (comparing to test/Foo$Bar in the normal mode), as getting qualified names from such internal names was trivial. But, because of IC, we needed to write class files to the disk, so our decompiler could find such "broken" classes and read it in a wrong way.